pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-powered-code-review-assistant"
version = "0.1.0"
description = "AI-powered code review assistant for style, complexity, and bug-risk detection"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Sarika Shirolkar" }]
dependencies = [
  "openai>=1.0.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "radon>=6.0.1",
  "ruff>=0.6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
code-review-assistant = "code_review_assistant.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["code_review_assistant*"]
